+++ src/components/UI/ImportExportWizards/index.ts
@@ -1,4 +1,3 @@
-export { ImportExportPage } from './ImportExportPage';
 export { ExportWizard } from './ExportWizard';
 export { ImportWizard } from './ImportWizard';
 export { ExportSessionsWizard } from './ExportSessionsWizard';
